+++ src/components/Input/Input.component.tsx
@@ -34,7 +34,7 @@ const inputComponentMap: SpecificInputComponentMap = {
 /**
  * Renders a specific input component based on the provided `type` prop.
  *
- * @version 0.0.4
+ * @version 0.0.5
  *
  * @param {GenericInputProps<TType>} props - The generic input props.
  * @returns {Nullable<JSX.Element>}